refactored to correct class names, consolidated global vars, correctly redrawing table view on update
