+++ VYNC/VYNC.xcdatamodeld/VYNC.xcdatamodel/contents
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <model userDefinedModelVersionIdentifier="" type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="6254" systemVersion="14B25" minimumToolsVersion="Xcode 4.3" macOSVersion="Automatic" iOSVersion="Automatic">
-    <entity name="UserX" representedClassName="VYNC.UserX" syncable="YES">
+    <entity name="User" representedClassName="VYNC.User" syncable="YES">
         <attribute name="id" optional="YES" attributeType="Integer 64" defaultValueString="0" syncable="YES"/>
         <attribute name="username" optional="YES" attributeType="String" syncable="YES"/>
     </entity>
-    <entity name="VideoMessageX" representedClassName="VYNC.VideoMessageX" syncable="YES">
+    <entity name="VideoMessage" representedClassName="VYNC.VideoMessage" syncable="YES">
         <attribute name="createdAt" optional="YES" attributeType="String" syncable="YES"/>
         <attribute name="id" optional="YES" attributeType="Integer 64" defaultValueString="0" syncable="YES"/>
         <attribute name="recipientId" optional="YES" attributeType="Integer 64" defaultValueString="0" syncable="YES"/>
@@ -14,7 +14,7 @@
         <attribute name="videoId" optional="YES" attributeType="String" syncable="YES"/>
     </entity>
     <elements>
-        <element name="UserX" positionX="-54" positionY="-9" width="128" height="75"/>
-        <element name="VideoMessageX" positionX="-54" positionY="9" width="128" height="148"/>
+        <element name="User" positionX="-54" positionY="-9" width="128" height="73"/>
+        <element name="VideoMessage" positionX="-54" positionY="9" width="128" height="148"/>
     </elements>
 </model>
